refactor out REPL functionality, fix module for completions and test it
